Initial port of contents
Performs an initial port of the project into the new repository. This
port includes two scripts used to (1) generate the secure boot binaries
for Intel and ARM architectures and (2) archive the binaries as .zip and
.tar.gz files for each of the architectures.

Adds auxillary files for performing PR checks and automated attachment
of the archive files to the release, if the workflow is being run due to
the creation of a tag.

Signed-off-by: Joey Vagedes <joeyvagedes@microsoft.com>